examples of launching browser in HandInteractionExamples
Add a button that launches the Edge browser. This demonstrates that the browser slate can be launched while your application continues running. The web page that is launched contains controls modeled after the Unity GUI controls in the examples, showing a button, checkbox, slider, dropdown, and edit box. This demonstrates the option of integrating with existing web tools or of creating simple form UI using HTML and shows what this kind of interaction would feel like.